+++ astropy/io/fits/fitsrec.py
@@ -190,6 +190,47 @@ class FITS_rec(np.recarray):
         self.formats = None
         return self
 
+    def __setstate__(self, state):
+        meta = state[-1]
+        column_state = state[-2]
+        state = state[:-2]
+
+        super(FITS_rec, self).__setstate__(state)
+
+        for attr, value in zip(meta, column_state):
+            setattr(self, attr, value)
+
+    def __reduce__(self):
+        """
+        Return a 3-tuple for pickling a FITS_rec. Use the super-class
+        functionality but then add in a tuple of FITS_rec-specific
+        values that get used in __setstate__.
+        """
+
+        reconst_func, reconst_func_args, state = super(FITS_rec, self).__reduce__()
+
+        # Define FITS_rec-specific attrs that get added to state
+        column_state = []
+        meta = []
+
+        for attrs in ['_convert', '_heapoffset', '_heapsize', '_nfields', '_gap',
+                      '_uint', 'names', 'formats', 'parnames', '_coldefs']:
+
+            try:
+                # _coldefs can be Delayed, and file objects cannot be
+                # picked, it needs to be deepcopied first
+                if attrs == '_coldefs':
+                    column_state.append(self._coldefs.__deepcopy__(None))
+                else:
+                    column_state.append(getattr(self, attrs))
+                meta.append(attrs)
+            except AttributeError:
+                pass
+
+        state = state + (column_state, meta)
+
+        return reconst_func, reconst_func_args, state
+
     def __array_finalize__(self, obj):
         if obj is None:
             return