+++ astropy/io/fits/tests/test_table.py
@@ -4,6 +4,7 @@ import numpy as np
 from numpy import char as chararray
 
 from ....extern.six.moves import range
+from ....extern.six.moves import cPickle as pickle
 from ....io import fits
 from ....tests.helper import pytest
 
@@ -2323,10 +2324,45 @@ class TestTableFunctions(FitsTestCase):
             for idx in range(1, 3):
                 assert comparerecords(new_hdul[idx].data, t2.data)
 
-
     def test_new_coldefs_with_invalid_seqence(self):
         """Test that a TypeError is raised when a ColDefs is instantiated with
         a sequence of non-Column objects.
         """
 
         pytest.raises(TypeError, fits.ColDefs, [1, 2, 3])
+
+    def test_pickle(self):
+        """
+        Regression test for https://github.com/astropy/astropy/issues/1597
+
+        Tests for pickling FITS_rec objects
+        """
+
+        # open existing FITS tables (images pickle by default, no test needed):
+        with fits.open(self.data('tb.fits')) as btb:
+            # Test column array is delayed and can pickle
+            assert isinstance(btb[1].columns._arrays[0], Delayed)
+
+            btb_pd = pickle.dumps(btb[1].data)
+            btb_pl = pickle.loads(btb_pd)
+
+            # It should not be delayed any more
+            assert not isinstance(btb[1].columns._arrays[0], Delayed)
+
+            assert comparerecords(btb_pl, btb[1].data)
+
+        with fits.open(self.data('ascii.fits')) as asc:
+            asc_pd = pickle.dumps(asc[1].data)
+            asc_pl = pickle.loads(asc_pd)
+            assert comparerecords(asc_pl, asc[1].data)
+
+        with fits.open(self.data('random_groups.fits')) as rgr:
+            rgr_pd = pickle.dumps(rgr[0].data)
+            rgr_pl = pickle.loads(rgr_pd)
+            assert comparerecords(rgr_pl, rgr[0].data)
+
+        with fits.open(self.data('zerowidth.fits')) as zwc:
+            # Doesn't pickle zero-width (_phanotm) column 'ORBPARM'
+            zwc_pd = pickle.dumps(zwc[2].data)
+            zwc_pl = pickle.loads(zwc_pd)
+            assert comparerecords(zwc_pl, zwc[2].data)